Groups duplicate fixtures
